LayoutTests/fast/events/page-visibility-resize-event-order-expected.txt

This test checks that resize event fires after visibilitychange event when a frame is resized while the page is hidden.
To manually test, minimize this window and restore the window. You should see a series of "PASS" below:

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.visibilityState is "visible"
PASS document.hidden is false
PASS document.visibilityState is "hidden"
PASS document.hidden is true
PASS document.visibilityState is "visible"
PASS document.hidden is false
PASS didReceiveResizeEvent is true
PASS successfullyParsed is true

TEST COMPLETE
